arr_insert and arr_get tested

+++ helper.c
@@ -51,7 +51,7 @@ uint32_t utole(uint64_t in) {
 }
 
 // Create new dynamically allocated file_t struct
-file_t* new_file_t(char* name, uint32_t offset, uint32_t length, int32_t index) {
+file_t* new_file_t(char* name, uint64_t offset, uint32_t length, int32_t index) {
 	file_t* f = salloc(sizeof(*f));
 	update_file_name(name, f);
 	update_file_offset(offset, f);
